feat: deprecate Digest adapter
The Digest adapter uses the legacy MD5 algorithm for hashing, per standard tooling such as the `htdigest` tool provided by the Apache web server.
While the specification has been updated to allow for SHA-256 and SHA-512, these would (a) require changes to how we request and validate credentials, and (b) changes to how the file is generated, and should likely have their own dedicated adapters.
Additionally, moving to one or both of those and immediately deprecating the existing functionality would break existing sites that were built with htdigest.
As such, this patch **deprecates** the Digest adapter, and notes the security concerns leading to that deprecation.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>